requirements.txt
Django>=4.0,<7.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.0
django-cors-headers>=4.0.0
django-ratelimit>=4.0.0
psycopg2-binary>=2.9.0

